refactor(side_menu_nav_drawer,item,header): rename locally-used helper widgets for drawer side menu
