+++ lib/shared/navigation/widgets/side_menu_nav_drawer/side_menu_nav_drawer_widget.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/material.dart';
 
-part 'create_drawer_item.dart';
-part 'create_header.dart';
+part 'drawer_header.dart';
+part 'drawer_item.dart';
